Add a Docker one-liner run example to the README
I had trouble running the provided Docker image on my Mac and my Raspberry Pi because it was only published in amd64 architecture. While we should probably publish the image in all architectures, this is a quick documentation fix to provide a one-liner for folks to run with Docker on any architecture using Caddy.